+++ docs/internal/development/retire-deprecated-workstation-join-config-data-model.md
@@ -38,15 +38,15 @@ that remains canonical.
 | Per-input guard | public config contract | Supported fan-in authoring surface on workstation inputs. Guard entries define child-completion and child-failure dependency behavior. | `interfaces.IOConfig.Guard`, `interfaces.WorkstationIOGuardConfig`, and generated `WorkstationIO.guards` | `pkg/interfaces/factory_config.go`, `libraries/agent-factory/api/openapi.yaml`, `pkg/config/openapi_factory_test.go` |
 | Workstation guard | public config contract | Workstation-level guard surface retained only for visit-count behavior. It must not carry parent-aware child fan-in. | `interfaces.FactoryWorkstationConfig.Guards` and generated `Workstation.guards` | `pkg/config/config_validator.go`, `pkg/config/config_mapper.go`, `docs/processes/agent-factory-development.md` |
 | Retired workstation join | removed config contract | Former workstation-level fan-in object that duplicated per-input guard behavior. It is no longer representable in generated types or active config structs. | Removed from `interfaces.FactoryWorkstationConfig`, OpenAPI schemas, and generated API models | cleanup analyzer report, `rg` US-007 inventory |
-| Dynamic fanout count | Petri-net runtime behavior | Count token and guard behavior used when a per-input child guard specifies `spawned_by`. | `pkg/config` mapper and Petri guard types | `pkg/config/config_mapper.go`, `pkg/config/config_mapper_test.go`, `pkg/replay/effective_config_test.go` |
+| Dynamic fanout count | Petri-net runtime behavior | Count token and guard behavior used when a per-input child guard specifies `spawned_by`. | `pkg/config` mapper and Petri guard types | `pkg/config/config_mapper.go`, `pkg/config/config_mapper_test.go`, `pkg/replay/configtests/effective_config_test.go` |
 | Parent-aware child failure | Petri-net runtime behavior | `any_child_failed` per-input guard route that observes failed children for the same parent context. | `pkg/config` mapper and Petri guard types | `pkg/config/config_mapper_test.go`, `tests/functional_test/config_driven_test.go`, `tests/stress/barrier_limits_test.go` |
 
 ## Identifiers
 
 | Identifier | Format | Producer | Consumer | Validation evidence |
 | --- | --- | --- | --- | --- |
-| `parent_input` / `parentInput` | input name string | config authors, OpenAPI clients, UI generated clients | config mapper, config validator, replay generated-factory conversion | `pkg/config/config_validator.go`, `pkg/config/openapi_factory_test.go`, `pkg/replay/effective_config_test.go` |
-| `spawned_by` / `spawnedBy` | workstation name string | config authors, generated-factory replay payloads | config mapper and validator for dynamic fanout count behavior | `pkg/config/config_mapper_test.go`, `pkg/config/config_validator_test.go`, `pkg/replay/effective_config_test.go` |
+| `parent_input` / `parentInput` | input name string | config authors, OpenAPI clients, UI generated clients | config mapper, config validator, replay generated-factory conversion | `pkg/config/config_validator.go`, `pkg/config/openapi_factory_test.go`, `pkg/replay/configtests/effective_config_test.go` |
+| `spawned_by` / `spawnedBy` | workstation name string | config authors, generated-factory replay payloads | config mapper and validator for dynamic fanout count behavior | `pkg/config/config_mapper_test.go`, `pkg/config/config_validator_test.go`, `pkg/replay/configtests/effective_config_test.go` |
 | `all_children_complete` | per-input guard type | config authors and generated Factory payloads | config mapper, Petri guard construction, replay conversion | `pkg/config/config_mapper_test.go`, `tests/functional_test/multi_input_guard_test.go` |
 | `any_child_failed` | per-input guard type | config authors and generated Factory payloads | config mapper, Petri guard construction, replay conversion | `pkg/config/config_mapper_test.go`, `tests/functional_test/config_driven_test.go`, `tests/stress/barrier_limits_test.go` |
 | `visit_count` | workstation guard type | config authors | workstation-level guard mapper and validator | `pkg/config/config_mapper.go`, `pkg/config/config_validator.go` |
@@ -61,7 +61,7 @@ states without introducing new runtime lifecycle values.
 
 | Config shape | Owner | Required fields | Defaults | Consumers | Evidence |
 | --- | --- | --- | --- | --- | --- |
-| `workstations[*].inputs[*].guards[]` | `interfaces.IOConfig.Guard` and OpenAPI `WorkstationIO.guards` | `type`, `work_type`, `state`; parent-aware guards also use `parent_input`; dynamic fanout uses `spawned_by` | No implicit join default; absent guards behave as normal input requirements | `pkg/config`, `pkg/replay`, generated UI/API clients | `api/openapi.yaml`, `pkg/config/config_mapper.go`, `pkg/replay/generated_factory_config_conversion.go` |
+| `workstations[*].inputs[*].guards[]` | `interfaces.IOConfig.Guard` and OpenAPI `WorkstationIO.guards` | `type`, `work_type`, `state`; parent-aware guards also use `parent_input`; dynamic fanout uses `spawned_by` | No implicit join default; absent guards behave as normal input requirements | `pkg/config`, `pkg/replay`, generated UI/API clients | `api/openapi.yaml`, `pkg/config/config_mapper.go`, `pkg/replay/generated_factory.go` |
 | `workstations[*].guards[]` | `interfaces.FactoryWorkstationConfig.Guards` and OpenAPI `Workstation.guards` | `type: visit_count` and visit-count fields | No parent-aware fan-in defaults | `pkg/config` mapper and validator | `pkg/config/config_validator.go`, `pkg/config/config_mapper.go` |
 | Removed workstation join object | none | none | none | none | cleanup analyzer report and US-007 zero-match inventory |
 
@@ -70,7 +70,7 @@ states without introducing new runtime lifecycle values.
 | Contract | Producer | Consumer | Allowed dependency direction | Error cases | Evidence |
 | --- | --- | --- | --- | --- | --- |
 | Public factory config fan-in | config authors, OpenAPI clients, UI generated clients | `pkg/config` loader, mapper, validator | Public schemas expose per-input guards; mapper translates those guards into Petri-net arcs. | Raw retired `join` payloads fail at the boundary because the generated/internal public shape cannot represent them. | `api/openapi.yaml`, `pkg/config/factory_config_mapping_test.go`, `pkg/config/config_mapper_test.go` |
-| Generated Factory replay fan-in | recorded generated Factory artifacts | `pkg/replay` hydration and config mapper | Replay uses explicit generated API to internal config mappers, preserving `WorkstationIO.guards`. | Generated artifacts cannot reintroduce `join`; unsupported guard shapes fail validation or mapping. | `pkg/replay/generated_factory_config_conversion.go`, `pkg/replay/effective_config_test.go` |
+| Generated Factory replay fan-in | recorded generated Factory artifacts | `pkg/replay` hydration and config mapper | Replay uses explicit generated API to internal config mappers, preserving `WorkstationIO.guards`. | Generated artifacts cannot reintroduce `join`; unsupported guard shapes fail validation or mapping. | `pkg/replay/generated_factory.go`, `pkg/replay/configtests/effective_config_test.go` |
 | Petri-net guard fan-in | config mapper | Petri runtime and functional/stress tests | Mapper creates observation/count arcs and parent-aware guards from per-input guard config. | Invalid `spawned_by`, parent input, or unsupported workstation-level child fan-in guard types fail validation. | `pkg/config/config_mapper.go`, `pkg/config/config_validator.go`, `tests/functional_test/config_driven_test.go`, `tests/stress/barrier_limits_test.go` |
 | Topology projection fan-in display | loaded runtime config | `pkg/factory/projections` and dashboard generated contracts | Projections read canonical runtime config guards and do not preserve retired join config. | Removed fields cannot appear in generated projection/API payloads. | `pkg/factory/projections/topology_projection.go`, generated API zero-match inventory |
