+++ docs/internal/development/workstation-runtime-config-data-model.md
@@ -41,7 +41,7 @@ runtime maps to one canonical loaded workstation config.
 | Config shape | Owner | Required fields | Defaults | Consumers | Evidence |
 | --- | --- | --- | --- | --- | --- |
 | `LoadedFactoryConfig.Workstations` | `pkg/config` | Workstation name key and one `interfaces.FactoryWorkstationConfig` value containing topology and resolved runtime fields. | Split `AGENTS.md` data is merged into the topology entry when present; topology-only script-worker dispatch defaults to `MODEL_WORKSTATION` at worker execution. | `pkg/service`, `pkg/workers`, `pkg/factory/projections`, `pkg/replay` | `pkg/config/runtime_config.go`, `pkg/workers/workstation_executor_test.go` |
-| `RUN_REQUEST.payload.factory.workstations` | `pkg/replay` and generated API schema | Generated Factory workstation array containing the replayable topology and runtime fields used by the recorded run. | Workstation stop handling serializes through one canonical ordered `stopWords` array; current artifacts must not contain legacy split workstation stop-word fields or maps. | `pkg/service`, `pkg/workers`, `pkg/factory/projections` | `pkg/replay/generated_factory.go`, `pkg/replay/generated_factory_runtime.go`, `pkg/replay/artifact_test.go` |
+| `RUN_REQUEST.payload.factory.workstations` | `pkg/replay` and generated API schema | Generated Factory workstation array containing the replayable topology and runtime fields used by the recorded run. | Workstation stop handling serializes through one canonical ordered `stopWords` array; current artifacts must not contain legacy split workstation stop-word fields or maps. | `pkg/service`, `pkg/workers`, `pkg/factory/projections` | `pkg/replay/generated_factory.go`, `pkg/replay/artifact_test.go` |
 | Split workstation layout | `pkg/config` | `factory.json` workstation entry and optional `workstations/<name>/AGENTS.md`. | Existing split files are preserved during expand when no inline runtime fields exist. | Config CLI, runtime loader, tests | `pkg/config/layout.go`, `pkg/config/runtime_config_test.go` |
 
 ## Inter-Package Contracts
