+++ docs/internal/processes/development-guide-relevant-files.md
@@ -136,6 +136,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `pkg/testutil/`, `pkg/testutil/harness/`, and `pkg/testutil/mockfactory/` | Shared test-helper package-size seam | Keep the root `pkg/testutil` package for exported helper APIs and same-package tests that need internals; move externally scoped behavior tests into focused subpackages such as `harness` and `mockfactory`, and keep internal contract inventory tests with their owning internal packages so shared test helpers do not accumulate unrelated contract coverage. |
 | `pkg/api/`, `pkg/api/contracttests/`, `pkg/api/boundarytests/`, and `pkg/api/servertests/` | API package-size test split seam | Keep root API tests for package-private handler, projection, and request-normalization helpers. Move OpenAPI/generated contract coverage into `contracttests`, cross-boundary public projection smoke tests into `boundarytests`, and black-box route or SSE behavior into `servertests` with small local HTTP/SSE helpers so the root API package stays under the file-count cap without exporting internals. |
 | `pkg/factory/subsystems/`, `pkg/factory/subsystems/dispatchertests/`, `pkg/factory/subsystems/circuitbreakertests/`, and `pkg/factory/subsystems/terminationtests/` | Factory subsystem package-size test split seam | Keep root subsystem tests for package-private transitioner reducers, mutation helpers, and integrated history-transitioner pipeline fixtures. Move exported constructor and `Execute` behavior coverage for dispatcher, no-op dispatcher, circuit breaker, termination, and tick-group ordering into focused external test packages so the root runtime subsystem package stays under the file-count cap without exporting internals. |
+| `pkg/replay/`, `pkg/replay/clocktests/`, `pkg/replay/configtests/`, and `pkg/replay/fixturetests/` | Replay package-size test split seam | Keep root replay tests for package-private event reducers, generated event-artifact normalization, and event-stream hydration helpers. Move exported-surface deterministic clock, generated-factory/config round-trip, and checked-in replay fixture safety coverage into focused external test packages so the root replay package stays under the file-count cap without exporting internals. |
 | `pkg/workcontent/generated.go` | Canonical generated work-content translation owner | Supported generated `factoryapi.WorkContent` to canonical `[]interfaces.WorkContentPart` translation should stay centralized here; API handlers, projection reconstruction, replay hydration, and future model-operation surfaces may keep boundary-local validation or error wording, but they should not fork the text, image, audio, JSON, or binary success-path translation logic again. |
 | `tests/functional/internal/support/api_server.go` | Shared live functional API-server lifecycle owner | Functional suites that need the same live API server, readiness wait, engine snapshot access, canonical event reads, or cancel-and-wait shutdown path should compose this helper instead of cloning `BuildFactoryService` plus `httptest.Server` startup in package-local `helpers_test.go` or `functional_server_test.go` wrappers; use the capture callbacks to preserve suite-local request or assertion conveniences, call the shared `GetFactoryEvents`/`Done`/`Stop` helpers instead of mirroring lifecycle fields locally, keep service-mode observability assertions behavioral by checking `Done()` stays open before explicit cancellation and closes after `Stop(t)`, and keep replay-only dashboard logic or runtime-only direct submission surfaces local to the owning suite. |
 | `tests/functional/internal/support/harness.go` | Shared functional harness-state helper owner | Functional suites that need reused marking token lookup, place token counts, or token-identity derivation against live harness snapshots should promote that logic here instead of keeping package-local token traversal helpers in runtime API or neighboring functional suites. |
@@ -545,7 +546,7 @@ This inventory records the checked-in files and directories that the maintainer 
 - Treat metadata-only split worker and workstation entries in expanded `factory.json` as incomplete without their paired body-only `AGENTS.md` files. `pkg/config/runtime_config.go` should fail load with explicit entity context instead of silently normalizing those entries to empty bodies.
 - Keep customer-path import/export regressions in `tests/functional/bootstrap_portability/api_export_import_e2e_smoke_test.go` observable at both seams: the API smoke should prove `GET /factory/~current` and `POST /factory` round-trip the canonical contract, and it should inspect the persisted named-factory directory to ensure imported `factory.json` stays thin while worker and workstation bodies remain body-only in split `AGENTS.md` files.
 - When `factoryapi.SubmitWorkRequest` contract fields change, sweep the direct unary submit helpers under `tests/functional/bootstrap_portability/` as well as the shared runtime API helpers; those bootstrap portability tests author `SubmitWorkRequest` payloads locally and can retain stale pointer or optional-field assumptions after generated contract updates.
-- When the exported workstation prompt field changes, keep `pkg/config/factory_config_mapping.go`, `pkg/replay/generated_factory_runtime.go`, and the generated-schema runtime smoke coverage aligned: the public `Factory` payload may serialize prompt content under `Workstation.body`, while replay and other runtime rebuild paths still need `PromptTemplate` restored from that canonical export field.
+- When the exported workstation prompt field changes, keep `pkg/config/factory_config_mapping.go`, `pkg/replay/generated_factory.go`, and the generated-schema runtime smoke coverage aligned: the public `Factory` payload may serialize prompt content under `Workstation.body`, while replay and other runtime rebuild paths still need `PromptTemplate` restored from that canonical export field.
 - Keep the PNG import/export boundary aligned with the canonical factory contract: PNG metadata should serialize the direct `Factory` fields plus `schemaVersion`, and the `/factory` API should receive the same `Factory` object without a wrapper or PNG-only fields.
 - Keep UI preview and drop state on the same canonical PNG contract object. Import/export helpers may carry `schemaVersion` alongside the `Factory`, but should not duplicate wrapper-era fields like `factoryName`, nested `factory`, or parallel envelope objects in local state.
 - Keep the handwritten UI factory-definition decoder and PNG importer strict to the generated contract. Retired factory-config aliases belong in historical fixtures only; the active browser boundary should reject them instead of normalizing them.
