+++ pkg/replay/configtests/effective_config_test.go
