+++ pkg/replay/configtests/generated_factory_test.go
@@ -120,15 +120,15 @@ func TestGeneratedFactoryFromLoadedConfig_EmbedsInlineDefinitionsWithoutConfigOn
 			"body":    "Run inline.",
 		}},
 		"workstations": []map[string]any{{
-			"id":             "execute-story-id",
-			"name":           "execute-story",
-			"worker":         "executor",
-			"inputs":         []map[string]string{{"workType": "story", "state": "init"}},
-			"outputs":        []map[string]string{{"workType": "story", "state": "complete"}},
-			"type":           "MODEL_WORKSTATION",
-			"body":           "Implement inline {{ .WorkID }}.",
-			"stopWords":      []string{"DONE"},
-			"limits":         map[string]any{"maxRetries": 2},
+			"id":        "execute-story-id",
+			"name":      "execute-story",
+			"worker":    "executor",
+			"inputs":    []map[string]string{{"workType": "story", "state": "init"}},
+			"outputs":   []map[string]string{{"workType": "story", "state": "complete"}},
+			"type":      "MODEL_WORKSTATION",
+			"body":      "Implement inline {{ .WorkID }}.",
+			"stopWords": []string{"DONE"},
+			"limits":    map[string]any{"maxRetries": 2},
 		}},
 	})
